Get rid of background-clip:text and PaintBehaviorForceBlackText.
http://blog.ericzhang.com/punch-through-text-masks-with-css-and-html5/
shows that you can achieve the similar effects using canvas. This
doesn't need to be built into the core engine. It's OK for this
to require more than a line of CSS to achieve.

Unfortunately, the test included this this patch doesn't actually
show that I didn't break anything because background-image
is broken. I got the test from
https://chromium.googlesource.com/chromium/blink/+/47e0d126e8d459b10176e3a91e47cf62ce0953a5%5E%21/#F4

R=eseidel@chromium.org

Review URL: https://codereview.chromium.org/756183004