Switch to SkyBinder
SkyBinder is a rewrite of TemplateBinding.sky that uses classes and weak maps
and doesn't add expandos to any of the DOM objects.

It has a few differences from TemplateBinding.sky:
- null doesn't create instances in repeats.
- One time bindings are removed ([[ expr ]]), we'll add them back later.
- The bind attribute is removed.

The system is also less generic and is specialized to be used when
creating SkyElement instances.

The commit history that created SkyBinder can be found at:
https://github.com/esprehn/sky-binder/commits/master

TBR=ojan@chromium.org

Review URL: https://codereview.chromium.org/821353003