Reduce latency of low-volume debugPrint output
If we haven't printed anything for a second, then reset the count of how
much we've printed.

I also reduced the amount of data to ~16kb until we pause because I saw
some corruption even pausing 1 second every ~32kb. (Numbers are
approximate because we're counting UTF-16 characters and some of the
characters in the output are multiple bytes in UTF-8.)