[Windows] Fix incorrect public C API argument type (#35106)
FlutterDesktopEngineGetTextureRegistrar is used to get the texture
registrar associated with an engine object and therefore should take a
FlutterDesktopEngineRef parameter.

This bug was initially identified and fixed by @knopp in:
https://github.com/flutter/engine/pull/27522

This patch replaces that one and adds the simplest possible unit test to
get it landed.

This also adds an initial unit test for the public Windows C API used to
implement the C++ client wrapper. In order to fully test this API, we'll
want to support test fixtures similar to what we do in the embedder API
tests. See: https://github.com/flutter/flutter/issues/87299

Fixes: https://github.com/flutter/flutter/issues/86617