[Impeller] Add Entity::RenderingMode enum. (#45845)
Resolves https://github.com/flutter/flutter/issues/134681.

For the various places where we track if a rendered entity is a subpass or not. After landing this, I think the next winning move is to make this an Entity field and remove the `SetRenderingMode` junk from the filter graph. That refactor is a bit more risky, so I plan to do that separately.