[Impeller] Add support for multi-rendering-backend fat shader archives. (#47278)
Existing shader archive ctors have been made private. Instead a factory is used to check to see if the blob is a single or multi-archive and do the right thing. This allows us to keep the existing code the same. But, when multi-archives are encountered, the right shaders libraries will be picked.

Creating multi-archives will be done in ImpellerC in an upcoming patch.

Towards https://github.com/flutter/flutter/issues/123741