[Impeller] started taking into account integer gaps in blur sigma, turned on linear filter (#48651)
A tiny tweak that should be better math since the sigma will scale based on the actual scale amount considering we have to scale down to discrete boundaries for the downsampled texture.  I also turned on linear sampling while I was at it to try to calm down the jitters.  I think it looks better at small sigmas.

test: This is covered by existing golden tests (when we turn this on).

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style