[Impeller] Add the KHR prefix to existing swapchain utilities. (#51295)
This is part of blowing up https://github.com/flutter/engine/pull/51213 into smaller commits.

Our current swapchain is based on the VK_KHR_swapchain extension. When this was the only swapchain that could be created, the KHR prefix was dropped. Since we are going to be having multiple swapchain types, add the KHR prefix and make room for other swapchains.

No change in functionality. Just renames classes and files.