[iOS][macOS] Eliminate use of bitcode_strip (#54254)
Our executables are no longer built with bitcode enabled and thus `bitcode_strip -r SOURCE -o DEST` is just copying the file in question to the output location.

Use of Bitcode was eliminated in Flutter in 2022. See linked issue for details.

This is a reland of https://github.com/flutter/engine/pull/54240 which was reverted in https://github.com/flutter/engine/pull/54250.

The the previous version was reverted because Python's `shutil.copyfile` doesn't set the unix permissions of the source file on the destination file. However, when overwriting an existing destination file, it preserves any permissions on that file. One can imagine how this might be problematic if you test the script by running before and after the change with the same output directory. `shutil.copy2` attempts to preserve source file metadata when writing the copy.

Issue: https://github.com/flutter/flutter/issues/107884

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style